+++ src/PlayGridManager.h
@@ -196,7 +196,7 @@ public:
      * @param duration An optional duration (0 means don't schedule a release)
      * @param delay A delay in ms counting from the beat
      */
-    Q_INVOKABLE void scheduleNote(unsigned char midiNote, unsigned char midiChannel, bool setOn = true, unsigned char velocity = 64, int duration = 0, int delay = 0);
+    Q_INVOKABLE void scheduleNote(unsigned char midiNote, unsigned char midiChannel, bool setOn = true, unsigned char velocity = 64, quint64 duration = 0, quint64 delay = 0);
 
     QObject *syncTimer() const;
     void setSyncTimer(QObject *syncTimer);
